+++ src/app/context/FinancialYearContext.tsx
@@ -1,6 +1,10 @@
 // context/FinancialYearContext.tsx
 "use client";
 
+import {
+  generateFinancialYears,
+  getCurrentFinancialYear,
+} from "@/utils/financialYearsHelperFunctions";
 import { createContext, useContext, useState, useEffect } from "react";
 
 type FinancialYearContextType = {
@@ -52,37 +56,6 @@ export function FinancialYearProvider({
 // Helper functions
 function validateFYFormat(year: string): boolean {
   return /^\d{4}-\d{4}$/.test(year);
-}
-
-function getCurrentFinancialYear(): string {
-  const today = new Date();
-  const year = today.getFullYear();
-  return today.getMonth() >= 3 ? `${year}-${year + 1}` : `${year - 1}-${year}`;
-}
-
-function generateFinancialYears(): string[] {
-  const currentFY = getCurrentFinancialYear();
-  const [startYearStr] = currentFY.split("-");
-  const currentStartYear = parseInt(startYearStr);
-
-  const years: string[] = [];
-
-  // Generate last 7 years
-  for (let i = 7; i > 0; i--) {
-    const year = currentStartYear - i;
-    years.push(`${year}-${year + 1}`);
-  }
-
-  // Include current year
-  years.push(currentFY);
-
-  // Generate next 2 years
-  for (let i = 1; i <= 2; i++) {
-    const year = currentStartYear + i;
-    years.push(`${year}-${year + 1}`);
-  }
-
-  return years;
 }
 
 export const useFinancialYear = () => {
